Updated ViewController
_Why_?

The project would not build...

In the chan.on("error") block we were referencing self.chat.text rather
than self.chatWindow.text.

_How_?

replaced the reference to self.chat.text with self.chatWindow.text

_Side Effects_?

seems to be operational, however I'm gettign kicked off of the phoenix
chat server after about 30 seconds. I wonder if this has to do with the
previous commit that relates to the heartbeat