Adds new features and upgrades compatability to Swift 4
This commit introduces the features of Presence and Logger (ability to
only show console messages on debug). A full rewrite based on the
current phoenix.js is underway. In the meantile, absolute credit for
the Presence feature goes to the Birdsong library. It has a great
Swift Channels impmentation and should be considered when looking for
an alternative. This also bumps a few items to make this compatable
with the new Swift 4 and iOS 11 releases.